fix: Correct serialization of `owner_id`
